refactor: move body tide Love/Shida numbers to `arguments`
