+++ doc/source/api_reference/arguments.rst
@@ -36,6 +36,8 @@ Calling Sequence
 
 .. autofunction:: pyTMD.arguments._constituent_parameters
 
+.. autofunction:: pyTMD.arguments._love_numbers
+
 .. autofunction:: pyTMD.arguments._parse_tide_potential_table
 
 .. autofunction:: pyTMD.arguments._to_doodson_number
